+++ ui/src/App.js
@@ -1,21 +1,17 @@
 import React, {Component} from 'react';
 import './css/App.css';
 import {BrowserRouter as Router, Route} from "react-router-dom";
-import {LoginForm} from "./components/login/LoginForm";
-import {QuestionsForm} from "./components/forms/QuestionsForm";
+import {QuestionsContainer} from "./components/container/QuestionsContainer";
 import '../node_modules/highlight.js/styles/github.css';
 
 class App extends Component {
 
     render() {
         return (
+            // TODO: about switch https://reacttraining.com/react-router/web/api/Switch
             <Router>
                 <div className="app">
-                    <Route exact path="/" component={QuestionsForm}/>
-                    <Route exact path="/questions" component={QuestionsForm}/>
-                    <Route exact path="/add" component={QuestionsForm}/>
-                    <Route exact path="/account" component={QuestionsForm}/>
-                    <Route path='/login' component={LoginForm}/>
+                    <Route exact path="/" component={QuestionsContainer}/>
                 </div>
             </Router>
         );
